requirements.txt
numpy==1.16.3
matplotlib==2.2.2
gym==0.14.0
Pillow==6.2.2
